% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterClik.R
\name{ClusterClik}
\alias{ClusterClik}
\title{ClusterClik}
\usage{
ClusterClik(matrice, R = 0.8, N = NULL, dist_method = 9)
}
\arguments{
\item{matrice}{une matrice ou un data.Frame}

\item{R}{une valeur entre 0 et 1 definissant un seuil de correlation}

\item{N}{une valeur entiere du nombre minimum d'objets dans un cluster}

\item{dist_method}{voir 'method' de la fonction dist.binary package ade4}
}
\value{
un vecteur contenant pour chaque objet la valeur de son cluster
}
\description{
ClusterClik
}
