% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhiCluster.R
\name{PhiCluster}
\alias{PhiCluster}
\title{PhiCluster}
\usage{
PhiCluster(matrice, R = 0, N = NULL)
}
\arguments{
\item{matrice}{une matrice ou un data.Frame}

\item{R}{une valeur entre -1 et 1 definissant un seuil de correlation}

\item{N}{une valeur entiere du nombre minimum d'objets dans un cluster}
}
\value{
un vecteur contenant pour chaque objet la valeur de son cluster
}
\description{
PhiCluster
}
